requirements.txt
# Teams Bot Core Dependencies
botbuilder-core==4.14.0
botbuilder-schema==4.14.0

# Azure Authentication & Graph API
msal==1.24.0

# Azure Services
azure-cognitiveservices-speech==1.32.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-mgmt-web==7.2.0
azure-mgmt-botservice==2.0.0
azure-mgmt-resource==23.0.0

# Web Framework & Async Support
aiohttp==3.8.5
flask==2.3.3

# Configuration Management
python-dotenv==1.0.0

# Audio Processing
pydub==0.25.1
pyaudio==0.2.11

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Development Tools
black==23.7.0
flake8==6.1.0
mypy==1.5.1
pylint==2.17.5

# Type Checking
types-requests==2.31.0

# Logging and Monitoring
azure-monitor-opentelemetry==1.0.0
opencensus-ext-azure==1.1.9

# Additional Dependencies
requests==2.31.0
pyyaml==6.0.1
setuptools>=65.0.0